requirements.txt
streamlit
pandas
datasets
google-generativeai
python-dotenv
google-cloud-storage
